requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq
tavily-python>=0.3.0
nest_asyncio>=1.6.0
fastapi
uvicorn
pydantic>=2.0.0
python-dotenv
markdown
fpdf2
scikit-learn
huggingface_hub
langchain-huggingface
sentence-transformers
